pyproject.toml
[project]
name = "personal-ai-employee"
version = "0.1.0"
description = "Digital FTE - Personal AI Employee (Silver Tier)"
requires-python = ">=3.13"
dependencies = [
    # Core
    "watchdog>=3.0.0",
    "python-dotenv>=1.0.0",

    # Backend API
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "pydantic>=2.0.0",

    # MCP Protocol
    "mcp>=1.0.0",

    # Gmail Integration
    "google-auth-oauthlib>=1.2.0",
    "google-api-python-client>=2.100.0",

    # WhatsApp (Playwright-based)
    "playwright>=1.40.0",

    # LinkedIn (API)
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
